move source generation into separate module allow to pass a Generator to NormalModule change NormalModule constructor to options object
:eggplant:
