+++ lib/JavascriptModulesPlugin.js
@@ -7,6 +7,7 @@
 const Parser = require("./Parser");
 const Template = require("./Template");
 const ConcatSource = require("webpack-sources").ConcatSource;
+const JavascriptGenerator = require("./JavascriptGenerator");
 
 class JavascriptModulesPlugin {
 	apply(compiler) {
@@ -21,6 +22,15 @@ class JavascriptModulesPlugin {
 			});
 			normalModuleFactory.hooks.createParser.for("javascript/esm").tap("JavascriptModulesPlugin", options => {
 				return new Parser(options, "module");
+			});
+			normalModuleFactory.hooks.createGenerator.for("javascript/auto").tap("JavascriptModulesPlugin", options => {
+				return new JavascriptGenerator(options);
+			});
+			normalModuleFactory.hooks.createGenerator.for("javascript/dynamic").tap("JavascriptModulesPlugin", options => {
+				return new JavascriptGenerator(options);
+			});
+			normalModuleFactory.hooks.createGenerator.for("javascript/esm").tap("JavascriptModulesPlugin", options => {
+				return new JavascriptGenerator(options);
 			});
 			compilation.mainTemplate.hooks.renderManifest.tap("JavascriptModulesPlugin", (result, options) => {
 				const chunk = options.chunk;
