+++ lib/JsonModulesPlugin.js
@@ -5,10 +5,7 @@
 "use strict";
 
 const JsonParser = require("./JsonParser");
-const ConcatSource = require("webpack-sources").ConcatSource;
-
-const stringifySafe = data => JSON.stringify(data)
-	.replace(/\u2028|\u2029/g, str => str === "\u2029" ? "\\u2029" : "\\u2028"); // invalid in JavaScript but valid JSON
+const JsonGenerator = require("./JsonGenerator");
 
 class JsonModulesPlugin {
 	apply(compiler) {
@@ -18,29 +15,8 @@ class JsonModulesPlugin {
 			normalModuleFactory.hooks.createParser.for("json").tap("JsonModulesPlugin", () => {
 				return new JsonParser();
 			});
-			compilation.moduleTemplates.javascript.hooks.content.tap("JsonModulesPlugin", (moduleSource, module) => {
-				if(module.type && module.type.startsWith("json")) {
-					const source = new ConcatSource();
-					const data = module.buildInfo.jsonData;
-					if(Array.isArray(module.buildMeta.providedExports) && !module.isUsed("default")) {
-						// Only some exports are used: We can optimize here, by only generating a part of the JSON
-						const reducedJson = {};
-						for(const exportName of module.buildMeta.providedExports) {
-							if(exportName === "default")
-								continue;
-							const used = module.isUsed(exportName);
-							if(used) {
-								reducedJson[used] = data[exportName];
-							}
-						}
-						source.add(`${module.moduleArgument}.exports = ${stringifySafe(reducedJson)};`);
-					} else {
-						source.add(`${module.moduleArgument}.exports = ${stringifySafe(data)};`);
-					}
-					return source;
-				} else {
-					return moduleSource;
-				}
+			normalModuleFactory.hooks.createGenerator.for("json").tap("JsonModulesPlugin", () => {
+				return new JsonGenerator();
 			});
 		});
 	}
