+++ lib/NormalModuleFactory.js
@@ -61,6 +61,8 @@ class NormalModuleFactory extends Tapable {
 			module: new SyncWaterfallHook(["module", "data"]),
 			createParser: new HookMap(() => new SyncBailHook(["parserOptions"])),
 			parser: new HookMap(() => new SyncHook(["parser", "parserOptions"])),
+			createGenerator: new HookMap(() => new SyncBailHook(["generatorOptions"])),
+			generator: new HookMap(() => new SyncHook(["generator", "generatorOptions"])),
 		};
 		this._pluginCompat.tap("NormalModuleFactory", options => {
 			switch(options.name) {
@@ -89,6 +91,7 @@ class NormalModuleFactory extends Tapable {
 		this.cachePredicate = typeof options.unsafeCache === "function" ? options.unsafeCache : Boolean.bind(null, options.unsafeCache);
 		this.context = context || "";
 		this.parserCache = Object.create(null);
+		this.generatorCache = Object.create(null);
 		this.hooks.factory.tap("NormalModuleFactory", () => (result, callback) => {
 			let resolver = this.hooks.resolver.call(null);
 
@@ -118,16 +121,7 @@ class NormalModuleFactory extends Tapable {
 							return callback(new Error("Empty dependency (no request)"));
 						}
 
-						createdModule = new NormalModule(
-							result.type,
-							result.request,
-							result.userRequest,
-							result.rawRequest,
-							result.loaders,
-							result.resource,
-							result.parser,
-							result.resolveOptions
-						);
+						createdModule = new NormalModule(result);
 					}
 
 					createdModule = this.hooks.module.call(createdModule, result);
@@ -254,6 +248,7 @@ class NormalModuleFactory extends Tapable {
 							settings,
 							type,
 							parser: this.getParser(type, settings.parser),
+							generator: this.getGenerator(type, settings.generator),
 							resolveOptions
 						});
 					});
@@ -347,6 +342,29 @@ class NormalModuleFactory extends Tapable {
 		}
 		this.hooks.parser.for(type).call(parser, parserOptions);
 		return parser;
+	}
+
+	getGenerator(type, generatorOptions) {
+		let ident = type;
+		if(generatorOptions) {
+			if(generatorOptions.ident)
+				ident = `${type}|${generatorOptions.ident}`;
+			else
+				ident = JSON.stringify([type, generatorOptions]);
+		}
+		if(ident in this.generatorCache) {
+			return this.generatorCache[ident];
+		}
+		return this.generatorCache[ident] = this.createGenerator(type, generatorOptions);
+	}
+
+	createGenerator(type, generatorOptions = {}) {
+		const generator = this.hooks.createGenerator.for(type).call(generatorOptions);
+		if(!generator) {
+			throw new Error(`No generator registered for ${type}`);
+		}
+		this.hooks.generator.for(type).call(generator, generatorOptions);
+		return generator;
 	}
 
 	getResolver(type, resolveOptions) {
