+++ lib/WebAssemblyModulesPlugin.js
@@ -5,6 +5,7 @@
 "use strict";
 
 const WebAssemblyParser = require("./WebAssemblyParser");
+const WebAssemblyGenerator = require("./WebAssemblyGenerator");
 const WebAssemblyImportDependency = require("./dependencies/WebAssemblyImportDependency");
 
 class WebAssemblyModulesPlugin {
@@ -16,6 +17,10 @@ class WebAssemblyModulesPlugin {
 
 			normalModuleFactory.hooks.createParser.for("webassembly/experimental").tap("WebAssemblyModulesPlugin", () => {
 				return new WebAssemblyParser();
+			});
+
+			normalModuleFactory.hooks.createGenerator.for("webassembly/experimental").tap("WebAssemblyModulesPlugin", () => {
+				return new WebAssemblyGenerator();
 			});
 
 			compilation.chunkTemplate.hooks.renderManifest.tap("WebAssemblyModulesPlugin", (result, options) => {
